feat: Add reflections from lesson 1
I'm listing this as a feature commit because this repo will only contain
my reflections on the course "Git and Github". There won't be any
program codes. Hence, my reflections will be taken as "program codes".

This is a huge commit. I had not learnt about `git commit' before this
juncture.